fix: resolve all build errors for ESP-IDF v5.4 + add build/simulation guide
- Fix RISC-V uint32_t format mismatch (-Werror=format) across examples 02-17
  by adding (unsigned) casts or PRIu32 macros
- Fix esp_task_wdt_init() API change (v4→v5 struct form) in 05_watchdog
- Add missing CMakeLists REQUIRES: esp_partition, esp_mm, spi_flash
- Remove standalone esp_task_wdt from 18_boot_diagnosis REQUIRES (merged into esp_system)
- Add CONFIG_ESPTOOLPY_FLASHSIZE_4MB to 21_core_dump sdkconfig.defaults
- Add sdkconfig.defaults for 22_gdb_jtag (FREERTOS_USE_TRACE_FACILITY)
- Add build_example.sh and run_all_qemu.sh scripts
- Add BUILD_AND_SIMULATION_GUIDE.md covering VS Code and Ubuntu CLI workflows

+++ examples/04_queues/main/main.c
@@ -18,6 +18,7 @@
  */
 #include <stdio.h>
 #include <stdlib.h>
+#include <inttypes.h>
 #include "freertos/FreeRTOS.h"
 #include "freertos/task.h"
 #include "freertos/queue.h"
@@ -88,7 +89,7 @@ static void producer_task(void *arg) {
          */
         if (xQueueSend(g_data_queue, &pkt, pdMS_TO_TICKS(50)) != pdTRUE) {
             dropped++;
-            ESP_LOGW(TAG, "[%s] Queue FULL — packet dropped! (total drops: %u)",
+            ESP_LOGW(TAG, "[%s] Queue FULL — packet dropped! (total drops: %" PRIu32 ")",
                      names[id], dropped);
         }
 
@@ -117,18 +118,18 @@ static void consumer_task(void *arg) {
             total++;
             received[rx.sensor_id]++;
 
-            printf("[Consumer] #%-4u  %s id=%u  val=%.2f %s  t=%u ms  queue_depth=%u\n",
+            printf("[Consumer] #%-4" PRIu32 "  %s id=%u  val=%.2f %s  t=%" PRIu32 " ms  queue_depth=%u\n",
                    total,
                    names[rx.sensor_id],
                    rx.sensor_id,
                    rx.value,
                    units[rx.sensor_id],
                    rx.timestamp_ms,
-                   uxQueueMessagesWaiting(g_data_queue));
+                   (unsigned)uxQueueMessagesWaiting(g_data_queue));
         } else {
             /* Timeout — no data for 2 s, print summary */
-            printf("[Consumer] === Summary after %u received ===\n", total);
-            printf("  Temp:  %u  Humid: %u  Press: %u\n",
+            printf("[Consumer] === Summary after %" PRIu32 " received ===\n", total);
+            printf("  Temp:  %" PRIu32 "  Humid: %" PRIu32 "  Press: %" PRIu32 "\n",
                    received[SENSOR_TEMP], received[SENSOR_HUMID], received[SENSOR_PRESS]);
         }
     }
